split generated controllers
